Move job list to project-template
This commit moves lists of our Tempest CI jobs into a project-template
and makes our project definition use that template. This way we enable
the list to be reused in our kuryr-tempest-plugin to always keep it in
sync.

Also the lower-constraints jobs are removed here and a template with
them is added instead.

Change-Id: I34e328322538127d9c7d0b1d8f9b8edf853b8ace