Ensure pools support don't break due to neutron-vif
This patch ensures multi-pool support is not broken by the fact
that neutron-vif requires admin rights to fully operate. In case
kuryr is running without admin rights, the neutron-vif related ports
will be removed instead of trying to put them back onto the pool, as
there is no access to the needed information to recover them.

Related-Bug: 1808506
Change-Id: I6eba3bddf4649b4f817dfd7b8463144c4ad69789