Drop Octavia providers supported protocols list
Currently, Kuryr LBaaS driver maintains a list of
supported protocols per Octavia provider.
This list is used for optimizing listener creation
flow (don't send a request to Octavia if the provider
doesn't support this protocol).

Since this list is actually an Octavia attribute, it will
be more robust to delete this list from Kuryr and handle
the corresponding exception in case specific protocol
is not supported by Octavia provider.

Change-Id: I94d6de873fc4d3bfafdfd591a15ba82fb4e38ec3
Closes-Bug: 1809049