Fixed variables error with 'public-subnet'
The current OpenStack Tacker project uses Kuryr-Kubernetes as a network.
When installing Kubernetes VIM through Devstack, the subnet name must be
changed according to the option.
However, currently, it is a fixed value so that it occurs error.
To prevent this, you need to modify the current fixed value as a variable.

Change-Id: I5c122adb6999d9761f97d52222999a2ac50eff39
Closes-Bug:#1812878
Related-Bug:#1812766