Switch to CentOS Amphora
Looks like Ubuntu's Amphora is broken, probably due to latest update of
ubuntu-minimal cloud image (which includes cloud-init version bump that
might be the culprit). This commit switches our gates to CentOS amp to
work the problem around until we fix the Ubuntu one.

Also the timeout for the jobs is raised to 3 hours because it seems like
building containers in the gate take ages.

Closes-Bug: 1822862
Change-Id: I3650aac01334e924ca64fcc4e79928141e010e5a