Fix SG rule creation when Pod is created after NP
If a pod is created right after a Network Policy is created,
and the pod is selected by the policy rules, it's possible that
while handling the NP creation event the pod still does not have
an IP assigned to it, causing a rule to be created without a
remote_ip defined and thus result on allow to/from everywhere SG rule.

This commit fixes the issue by skipping the SG rule creation if
the pod does not have an IP assigned, and let the creation
happen after the pod is running.

Change-Id: I27cc52efea18342bc29ebe02343a98b5a401831a
Closes-Bug: 1829389