Fix CRD update on SG rule creation
The CRD is not updated with the created SG rule
when a Network Policy that contains more than one selector
in the ingress/egress rules and a pod that matches any selector
that is not the first are created.
This commit fixes the issue by ensuring the flag that indicates
when the CRD needs update is not overwritten when handling each
selector.

Change-Id: I02c9b9929a3156d9d40e5aeefef5456f6befa0ab
Closes-Bug: 1829041