Fix Multiple SG rules created when Pod creation follows a NP creation
If a pod is created right after a Network Policy that allows traffic
from a pod on a port is created, and the pod is selected by
the policy rules, it's possible that while handling the NP creation event
the pod still does not have an IP assigned to it, causing a rule to be
created without a remote_ip defined and thus result on allow from
everywhere in the specific port SG rule.

Prior to creating a SG rule, the existence of the cidr associated
to the k8s resources needs to be validated.

Change-Id: Id8692d06de4d3dbdfc6a719d4502ecc9aac933af
Closes-Bug: 1829998