Fix Pod creation do not triggers affected SVC SG update
The LBaaS SG is not updated when a pod that is referenced by a
service has a network policy that restrict/accept traffic from
pods in different namespaces. For example: pod-A on namespace-A is
created and according to a network policy should affects the traffic
to pod-B on namespace-B, and consequently to the service that has
pod B as a backend pod.

Change-Id: Idb3365e70e66dd8544f5b97aece312bc18c57ee5
Closes-Bug: 1830598