Improve behaviour of VIFHandler
Previously VIFHandler run through all the vifs
waiting it to be active. If neutron ports was not
active then exception was raised and new pod state
was not written. Thus each retry VIFHandler should
run through all the vifs even if appropriate neutron
port is active. Check for each neutron port requires
request to neutron that is time consuming process.

This commit allows to update pod state even if state
of single port was changed regardless exceptions for
other ports.

Change-Id: I6f74831831e33c7f1bdbc04df78601f9522d9212
Signed-off-by: Danil Golov <d.golov@samsung.com>