Fix lBaaS sg rules corresponding to the NP rules are being deleted
The LBaaS security group rules created to enforce the Network Policy
applied on the backend pods are being deleted, because are considered
to be default lbaas rule, when they are not.
This commit fixes the issue by adding a new factor to define whether
a lbaas rule is default or not. It checks if the description of the
rule contains a part of the description added to the rules created
when parsing the network policy sg rules, if it has than it should
not be deleted.

Change-Id: Iebbbd75990fe626cede00440121a9d2c2f3efaa0
Closes-Bug: 1832010