Fix fail to recreate namespace when previous KuryrNet CRD is not deleted
When a namespace/project is recreated right after its deletion,
it's possible that the previous kuryrNet CRD is not yet removed,
and Kuryr will try to create another CRD, resulting in k8s conflict
exception. In the meantime, other events handling resources under
the same namespace in question could be inquiring for an annotation
on the namespace, that was not annotated due to the conflict error,
causing the controller to restart.

This commit fixes the issue by adding an annotation to the KuryrNet
CRD with the respective namespace uid, which allows to check
if the existing CRD corresponds to the namespace.

Change-Id: I85d903902b516bfd109078705860540d731886bc
Closes-Bug: 1833712