Fix adding pods with host networking to svc in L2 mode
If L2 mode is used for services, there is a need for stating the
subnet the pods belong to when adding them as members of the load
balancer. If these pods have host networking, then their IPs will
not be neither in the subnet CIDR range for the default pod subnet,
nor on the namespace subnet CIDR if namespace subnet driver is used.

To overcome this, we assume that in case the pod IP is not within
those ranges it means it is running on host networking mode. And
thus the worker_node_subnet CIDR range should be used instead.

Change-Id: I4ce376956b66047069b6544d36f5c3d47fa38b01
Closes-Bug: 1834605
(cherry picked from commit 0345cd86c801f635e7dcaccc3768b28e3b3e8439)