Speed up pools prepopulation on namespace creation
If the amount of nodes is large, prepopulating the pools on each
node when a namespace is created may take a considerable amount
of time. To limit the impact this patch ensures the initial
prepopulation is not as large as the normal ones that happen node
by node depending on the needs.

Change-Id: I2f3c18a64cdffa71fb1faacdd34f6972319164e9