Fix for k8s client annotation
This commit changes the behaviour of k8s client
while annotating. K8s client should not throw exceptions
if there are conflicts in data and resourceVersion
is different. It should resolve conflicts properly
and try to patch annotations once again changing
resourceVersion.

If data is the same in process of annotating, then
k8s client should not annotate many times.

Change-Id: I7d981b59ecdae7fe434fe580f2f342cc4621198e
Signed-off-by: Danil Golov <d.golov@samsung.com>