Fix wrong exception handled on listener creation
In case a bad request error occur when ensuring the
neutron resource is provisioned, we are not reraising
the exception to properly handle it nor we're catching
the right exception, instead a resource not ready exception
is raised when it timeout, causing the controller to continously
restart.

This commit fixes the issue by raising the bad request
exception and catching it.

Closes-Bug: 1836891

Change-Id: I325c56ce41380657d2e0e2eb225ba15b4645b7de