Ensure leftover ports are deleted upon kuryr-controller restart
This patch ensures that no leftover ports are left behind. It may
happen that the kuryr-controller crashes before either attaching
the subports to the trunk or before tagging them. This patch checks
the existing ports in properly created (and tagged) networks and
deletes the ones that did not complete their creation process, i.e.,
Device Owner set to compute:kuryr and status DOWN with:
- Not properly tagged if tags are used
- No binding information in the non-nested case

Closes-Bug: 1839062
Change-Id: Ie1b80082627bc49fc3e396333dedbf79e22e63c4