Ensure SG is only deleted in case it's present in the KuryrNet CRD
In case Namespace Isolation is not enforced no SG is created and
added to the KuryrNet CRD. Right now, when an exception occurred
on the KuryrNet CRD addition, we try to delete a SG that was not
created.
This commit fixes the issue by checking for the existence of the
sgID in the CRD pior to deleting it.

Change-Id: I4ad3886398138d9a4a8e483b66b3573200002a0c
Closes-Bug: 1841765