Skipping sg rule deletion for pods without annotations
This patch set skips the error when NPs are enabled and a pod that
has not yet been annotated by kuryr is deleted (for instance a pod
in pending status). The information about the IP is found on the
annotations and as that is not already there no rules would have
been created. Thus no need to remove them.

Change-Id: I850dc2d66fac2d7b44267382062093b19e5cd442
Closes-Bug: 1842099