Timeout connections when watching K8s API
Turns out that without `timeout` parameter, requests library is unable
to see connection interruption if those were "silent" (i.e. firewall
blocked the connection for a period long enough for K8s API to drop the
connection). This can be fixed by adding the `timeout` parameter to the
`requests.get()` invocation in `watch()` and retrying on receiving
`requests.ReadTimeout` exception.

This commit does so by setting the connection and read timeouts to 30
and 60 respectively and implementing the retry logic using last seen
resourceVersion to restart the watching connection.

Change-Id: I19df95aaae404b5a936c469085b1f80f9eb99e2e
Closes-Bug: 1842689