Ensure LBaaS state is updated upon SVC deletion
When a SVC is recreated all the LBaaS resources are first deleted,
which requires the LBaaS state obj to be updated accordingly, then
created. Right now, when a LBaaS deletion happens the values for
members, pools and listeners are not updated on the lbaas state obj.

This commit fixes the issue by ensuring the lbaas state object is
updated.

Closes-Bug: 1843142
Change-Id: Ibe72cb3ebd3d2a861684e7e8ac5a310fe8060aa9