Ensure pod IP is retrieved from annotation
We are retrieving the pod IP from a nonexistent annotation
field, when should be from the annotations in the pod metadata.
Also, when this information is not present we keep retrying with
the same pod object, which doesn't contains an annotation
causing the controller restart.

Closes-Bug: 1845635
Change-Id: I04095a7d36219175c09d4503f2347ab035d3e242
(cherry picked from commit 9dfc5a1dd36e7c1691f5fbcfe9ec506f8dc5d2b9)