Ensure no LBaaS SG update is triggered for SVCs without selectors and ports
When a Network Policy is enforced we shouldn't try to update the
SG of a LBaaS that would map to a SVC without selector, as this
kind of SVC is not wired by Kuryr. Also, we shouldn't try to update
the LBaaS SG when no ports are defined in the SVC spec.

Closes-Bug: 1845917
Change-Id: I94a288f2b66bd2444d177931f509e1b6ef250235
(cherry picked from commit 68145b9b589f27b2db0130e6671c54c15f844bc6)