Avoid namespace deletion error if processing a duplicated event
In case of a kuryr-controller restart, if the namespace deletion
event was already executed and it is retriggered, it can make the
kuryr-controller fail as the kuryrnet CRD would have been already
deleted and we cannot obtain the subnet id from it. In that case
the event just need to be skip.

Change-Id: I14fcf820d94319541dd7bf13893e82cf0373969c
Closes-Bug: 1847453
(cherry picked from commit efae1f5211e3907ff58853155b785f0e8fb34e9b)