Fix CNI_COMMAND=VERSION case
In case of VERSION command k8s v1.16 and higher sends
only {"cniVersion": "0.4.0"} to the stdio, but kuryr-cni
expected cni-conf with kuryr_conf field.

Also k8s expects json in the stdio, but logs was also there.
So need to strictly redirect it into stderr in this case.

It's a work around, because maintanable way here is to use
golang's version based on containernetworking library.

Change-Id: I9095ced11f78eb1ae2890883fa6da4df74213480
Signed-off-by: Alexey Perevalov <a.perevalov@samsung.com>