Ensure Network Policy handles egress traffic to a SVC
We're not taking into account the case of a Network Policy
with an egress rule to a pod that contains a Service sitting
in front of it. Right now, only an egress rule to the matched
pod is created, when one for the matched SVC is also required.

Related-Bug: 1849139
Change-Id: I9830f30ba1fde3e5ec1a98fcbca22af992dd1bec
(cherry picked from commit db1b24fcf627e00ca7a541164538bdfe860ddf2c)