Protection from subport with wrong ACTIVE status
On high spikes, it seems that neutron may fail to properly update
the status of subports when being detached from the trunks. This
leads to ports with ACTIVE status while they are not attached
anywhere. Under such circunstance Kuryr was not able to clean up
the network associated to the deleted namespace as the port
is not considered since it is neither DOWN nor part of a trunk.
This patch adds protection from this.

Change-Id: I50e8a4117fc261ef67803dc114638acf378e2ba3