Prevent Kuryr restart due to connection issue
In case the connection to Keystone is broken Kuryr controller is instantly
restarted. This may result in leftovers loadbalancers as a Kubernetes
service deletion can happen while the lbaas is still in PENDING_CREATE
and if the controller is restarted, the deletion event will be gone.
In this commit we give Kuryr a chance to retry the event before restarting.

Change-Id: Ie2cacc86e745bbd7da7254f7adfd9b41370936ab
Closes-bug: 1853460