Add requests.ConnectionError to watcher retries
Apparently with switch to Python 3, requests library raises a different
exception when read time out occurrs and it's quite general
requests.ConnectionError. This causes tracebacks like [1] to show up in
logs.

This is pretty scary for the user, so this commit suppresses those exceptions
and adds them to the list of exceptions K8sClient restarts watching.

[1] http://paste.openstack.org/show/786574/

Change-Id: I09f3803edffbe2ea5f1913238adc23212a323100