Handle not found subport on Trunk
When a namespace is created-deleted-created, the last namespace
creation event might notices that a KuryrNet CRD still exists,
as it can take a while for the previous deletion event to complete
and will trigger again another deletion causing NotFound SubPort
on Trunk error. This commit handles the exception by ignoring
the port deletion.

Change-Id: Id0683a00b2c90529e0106296b5f157cf6885388b
Closes-bug: 1854096